DetailsModal, OnlyLecturer, StudentManagement, CourseManagement - Overview

+++ src/pages/Logout/index.js
@@ -1,30 +1,52 @@
 import { useNavigate } from 'react-router-dom';
 import { Loader } from '../../ui';
 import { Container } from './styles';
-import { useAuth, useNotify } from '../../hooks';
-import { useEffect } from 'react';
-import { useQueryClient } from '@tanstack/react-query';
-import { Approutes } from '../../constants';
+import { useAxios, useNotify } from '../../hooks';
+import { Approutes, Appurls } from '../../constants';
+import { useEffect, useRef } from 'react';
+import { useMutation } from '@tanstack/react-query';
+import { CSRFToken } from '../../components';
 
 const Logout = () => {
+	const axios = useAxios();
+	const isFirstTime = useRef(true);
+
 	const notify = useNotify();
 	const navigate = useNavigate();
-	const { setAuthTokens, setUser } = useAuth();
-	const queryClient = useQueryClient();
+
+	const { mutate: logout } = useMutation({ mutationFn: () => axios.post(Appurls.auth.logout) });
 
 	useEffect(() => {
-		setUser(null);
-		setAuthTokens(null);
+		if (!isFirstTime.current) return;
+		isFirstTime.current = false;
 
-		queryClient.clear();
-		localStorage.clear();
-
-		navigate(Approutes.auth.login);
-		notify({ message: "You've been logged out successfully", status: 'success', toastOptions: { toastId: 'logged_out' } });
-	}, [queryClient, setAuthTokens, setUser, navigate, notify]);
+		logout(
+			{},
+			{
+				onSuccess: () => {
+					localStorage.clear();
+					notify({
+						message: 'You have been logged out',
+						status: 'success',
+						toastOptions: { toastId: 'logout_success' },
+					});
+					navigate(Approutes.auth.login);
+				},
+				onError: () => {
+					notify({
+						message: 'Something went wrong while trying to log you out',
+						status: 'error',
+						toastOptions: { toastId: 'logout_error' },
+					});
+					navigate(-1);
+				},
+			}
+		);
+	}, [logout, navigate, notify]);
 
 	return (
 		<Container>
+			<CSRFToken />
 			<Loader />
 		</Container>
 	);
